+++ bin/build.sh
@@ -36,5 +36,11 @@ cp node_modules/typeahead.js/dist/typeahead.bundle.min.js defaults/public/js/ove
 # typeahead.js-bootstrap3.less (to make typeahead fit with bootstrap styles)
 cp node_modules/typeahead.js-bootstrap3.less/typeahead.css defaults/public/css/overlook/css/
 
+# Trumbowyg HTML editor (http://alex-d.github.io/Trumbowyg/)
+mkdir -p defaults/public/css/overlook/svg
+cp node_modules/trumbowyg/dist/ui/trumbowyg.min.css defaults/public/css/overlook/css/
+cp node_modules/trumbowyg/dist/ui/icons.svg defaults/public/css/overlook/svg/trumbowyg-icons.svg
+cp node_modules/trumbowyg/dist/trumbowyg.min.js defaults/public/js/overlook/libraries/
+
 # finished
 echo "Done"