+++ defaults/public/js/overlook/htmlEdit.js
@@ -1,0 +1,4 @@
+$(document).ready(function() {
+    $.trumbowyg.svgPath = '/css/overlook/svg/trumbowyg-icons.svg';
+    $('.html-edit').trumbowyg();
+});