+++ changelog.md
@@ -678,3 +678,7 @@ No changes. Erroneous version bump.
 * Fix: `lib/forms/displayers` URL displayer uses `target="_blank"`
 * Fix: `textboxSpaced` form data sanitizer handle line breaks properly
 * Field format for Twitter handles
+
+## Next
+
+* HTML editor form fields using `trumbowyg`